third_party/WebKit/LayoutTests/fast/canvas-api/canvas-2d-imageData-create-nonfinite.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<canvas id="canvas" width=600 height=300>
<script>
test(function(t) {
    var canvas = document.getElementById('canvas');
    var ctx = canvas.getContext('2d');

    try {
      ctx.createImageData(Infinity, Infinity);
    } catch (ex) {
      assert_equals(ex.name, "IndexSizeError");
    }
    try {
      ctx.createImageData(Infinity, 10);
    } catch (ex) {
      assert_equals(ex.name, "IndexSizeError");
    }
    try {
      ctx.createImageData(-Infinity, 10);
    } catch (ex) {
      assert_equals(ex.name, "IndexSizeError");
    }
    try {
      ctx.createImageData(10, Infinity);
    } catch (ex) {
      assert_equals(ex.name, "IndexSizeError");
    }
    try {
      ctx.createImageData(10, -Infinity);
    } catch (ex) {
      assert_equals(ex.name, "IndexSizeError");
    }
    try {
      ctx.createImageData(NaN, 10);
    } catch (ex) {
      assert_equals(ex.name, "IndexSizeError");
    }
    try {
      ctx.createImageData(10, NaN);
    } catch (ex) {
      assert_equals(ex.name, "IndexSizeError");
    }
}, 'Test the argument bounds of canvas createImageData.');
</script> 
